Set assigned_by when assigning case to attorney
**Why**: When viewing a case, or fetching it via the API, we want to
display by whom it was assigned, which corresponds to the `demdusr`
field in VACOLS. When we first assign a case to an attorney in Caseflow,
we were only updating the `adding_user` (`deadusr` in VACOLS), which
meant that when viewing the case, the `assigned_by` attribute was `nil`.

To fix this, we make sure to set `modifying_user` as well when assigning
a case to an attorney.