Add bummr and derailed_benchmarks dev gems
**Why**: bummr is very useful for updating multiple gems at once using
separate commits. derailed_benchmarks is great for finding memory and
other performance issues.

The reason for pointing at the master branch for bummr is that there is
a bug in version 0.4.0 and the fix that is on master hasn't been
released yet. See: https://github.com/lpender/bummr/issues/58

References:
https://github.com/lpender/bummr
https://github.com/schneems/derailed_benchmarks