Update capybara from 2.18.0 to 3.15.0
One of the major new changes is that when comparing text content on the
page, Capybara no longer fully normalizes whitespace, instead returning
text as close as possible to "as rendered". For example, if there is a
new line between text, or if the text in the spec has 2 whitespaces but
the rendered text only has one whitespace, the test will fail.

To make these tests pass, the spec needs to include the newline in the
expectation. For text that includes extra whitespace, we can either
remove the extra whitespace from the text in the spec, or if what we're
really testing is that the rendered text removes extra whitespace, we
can call `.squeeze(" ")` on the text in the expectation.

See the changes in this PR for examples of what expectations should look
like.

Because the Capybara API changed in 3.x, we also had to update our
sniffybara gem to match the new API.