Handle when VSO employee can't access file number
**Why**: The controller design was resulting in a double render error
`AbstractController::DoubleRenderError` in the scenario where a VSO
employee doesn't have access to the file number. This scenario was not
tested.

This PR adds tests for that scenario, and others that weren't tested,
and also cleans up the controller by moving the business logic in
separate classes.

Fixes Sentry errors such as:
https://sentry.ds.va.gov/department-of-veterans-affairs/caseflow/issues/3663/events/241077/