Fix flakey test
contention_reference_id should be unique across all RequestIssue objects but we were sharing
it. Because some test data is created for all tests (even if it is not used) this was causing
non-deterministic behaviour when calling source_request_issues.first because the order of that
array is not predictable. Unique contention_reference_id ensures that source_request_issues
is always an array of length 1.