Don't pass in large files as ActiveJob arguments
**Why**: Amazon SQS limits the size of arguments passed to ActiveJob.
In #9917, we passed in the file to the job in order to fix the issue
where the background job didn't have the file in memory. That was not
the right solution.

The way to fix the issue with accessing the file attribute (that doesn't
get saved to the DB) is to separate the S3 file storage from the rest of
the process. To initiate the storage, we have to go through the
ActiveRecord model. If we call `cache_file` on the document object
created in `PrepareDocumentUploadToVbms`, then the document will have
the `file` attribute attached to it, and we can pass it to
`UploadDocumentToVbms` which will have access to it as well.