Enqueue case distribution job from controller (#10003)
**Why**: By enqueuing the background job via an `after_commit` callback
on the model, it makes it hard to test `Distribution#distribute!` in
isolation. It can also cause an infinite loop intermittently in specs
that both create a new Distribution object and call `#distribute!`
(which the job calls as well).

Instead of tightly coupling the job with the model, we only call the job
in the context in which it is used, which is in the controller
(triggered by a user clicking the "Request more cases" button).