Route back button to correct place after dispatch
**Why**: If a judge edited or deleted an issue during the dispatch
checkout of a legacy appeal, and then clicked the browser back button
after the checkout, they would end up on the dispositions page instead
of the appeal details page. If they clicked on the Caseflow back button
on the dispositions page, it would try to take them to the special
issues page, which caused an exception to be sent to Sentry because the
judge no longer has permissions to edit the special issues for this
appeal.

This only happened when editing issues because the "Add Issue" and
"Edit Issue" links on the Dispositions page were React `Link`s, which
by default add a new entry to the `history` stack. By passing in the
`replace` property, we allow the current entry in the history stack to
be replaced. This required updating the `caseflow-frontend-toolkit`
dependency to support the `replace` property. Huge thanks to Hunter
Schallhorn for figuring this out and fixing it, and to Anya Roltsch for
finding the steps to reproduce.

Resolves #9899