Fixing VSO method when there is non POA (#10256)
Resolves #10234

### Description
When a Veteran has no POA, the participant_id is nil. In this case we look up any VSOs in our database with a nil POA. For some reason we've recently added a few. Looks like some field VSOs may not have a participant id. Regardless we don't want to pull in any VSO if the Veteran doesn't have representation.