Fix Sentry alert issue of checking if value is defined before checking its length (#10283)
Resolves #10277 

There was a Sentry alert for trying to check the length of a value that was undefined, which throws an error. Add a simple check to make sure the value is defined before checking its length. The fact that this situation was hit was probably just a strange edge case (only one user experienced this once), so this is just a precaution.

### Description
Please explain the changes you made here.

### Acceptance Criteria
- [ ] Code compiles correctly

### Testing Plan
1. Go to ...