Removing asyncable from record_synced_by_job (#10244)
Remove Asyncable from record_synced_by_job. We inherited from Asyncable to try and use the same patterns as other jobs. However, it was overkill for this job since it already has a natural retry mechanism. If the job fails, it will retry an hour later. All we really need is a `processed_at` time.