only search for select in the modal body (#10501)
### Description
This test was flaking on larger monitors when the select control on the page behind the modal was still visible and therefore included in `find_all(".Select-control")`. The changes restrict the search for the select control to the modal body, which is where the test is concerned with it appearing.