Fix double counting of cases in attorney queues (#10534)
Before | After
--- | ---
![image](https://user-images.githubusercontent.com/32683958/56690430-6ca98300-66ab-11e9-8929-b7da26d9e13d.png) | ![image](https://user-images.githubusercontent.com/32683958/56690435-6fa47380-66ab-11e9-9be5-e778383b0759.png)


The attorney queue shows `ColocatedTask`s that the attorney created so that the attorney still knows that the case they are assigned is being worked on. This makes sense for `LegacyAppeal`s because the location code changes to `CASEFLOW` and the `AttorneyLegacyTask` goes away as a result, but for `Appeal`s we will double count the appeals because the parent task of the `ColocatedTask` will not go away (the `AttorneyTask` will stay around).