Fix ActionController::ParameterMissing bugs (#10622)
**Why**: When we disabled parameter wrapping in #10508, we affected
a couple of scenarios that weren't covered by feature specs. When
parameter wrapping is enabled, if a required parameter is not passed
in the JSON payload, Rails will automatically include it based on the
controller name. For example, if `HearingsController` defines
`params.require(:hearing).permit(:prepped)`, but the payload only
passes in an attribute, such as `{ prepped: true }`, Rails will
automatically include the `hearing` key based on the name of the
controller.

So far, based on Sentry alerts, we have identified two React views that
were not passing in the required parameter, and therefore started
failing after we disabled parameter wrapping.