Handle more error messages from VA.gov address validator API (#10668)
Resolves #10387

### Description

The address validator API was returning that an address was valid, but mapped to multiple locations. We weren't handling this error and the API returned a geocode of `(0.0, 0.0)`, which was leading to incorrectly calculated distances.

This PR adds some additional error handling to handle when the address validator returns an error in the `messages` field (see https://github.com/department-of-veterans-affairs/caseflow/issues/7507#issuecomment-488751114). The zip code fallback is used when the VA.gov API returns an address validation error that is reported in the messages field.

This PR also refactors the error handling in `va_dot_gov_address_validator.rb` to use exception handling as opposed to checking the first error message, which can be unreliable (may not have seen this be an issue in practice).