Add common component for AMA issues (#10284)
Resolves #8591

### Description
Create a shareable component for AMA issues. To be used in Case Details page and Add Decisions flow.

The way I decided to build this is by separating the display of issues and decisions that go with each issue. This choice to separate them has some pros and cons:

**Pros**: Decisions don't seem to be shown in every situation where issues are, so keeping them separate makes it easier to only use the component for issues when that's all that's needed. It also makes the issue component more flexible to show different components as children instead of just the one view of decisions.
**Cons**: It means a bit more manual configuration when using both together.

I think the trade-off to favor flexibility over reducing configuration is worthwhile, especially given the brief of this ticket, which is to create a shareable component between all apps that display issues, but open to thoughts here!

How this will now look in the Case Details page:
<img width="923" alt="Screen Shot 2019-04-04 at 2 39 01 PM" src="https://user-images.githubusercontent.com/25183456/55581505-0bfdda80-56eb-11e9-9731-233d94300a26.png">

How this will now look in the Add Decisions flow:
<img width="1204" alt="Screen Shot 2019-04-04 at 2 38 49 PM" src="https://user-images.githubusercontent.com/25183456/55581511-10c28e80-56eb-11e9-9367-f8955ed190dd.png">

<img width="901" alt="Screen Shot 2019-04-04 at 2 39 26 PM" src="https://user-images.githubusercontent.com/25183456/55581521-16b86f80-56eb-11e9-9c6a-afbbd5292382.png">

### Acceptance Criteria
- [ ] Code compiles correctly

### Testing Plan
1. Go to ...