Update document series ids with 1 SQL statement (#10789)
**Why**: One of our slowest transactions is
`Reader::DocumentsController#index` due to the call to `annotations`,
which triggers an update of hundreds of documents via
`AddSeriesIdToDocuments`. Instead of using a separate DB call for
each document, we can batch update the documents with a single DB
call. This is possible thanks to the `activerecord-import` gem which
takes advantage of some features only available in Postgres 9.5+, which
was the main reason I have been pushing for us to upgrade our Postgres
version.

I also took this opporturity to do some refactoring.