Inject docket number column into TaskTable instance in OrganizationQueue. (#10735)
Partially resolves #10619, which calls for injecting columns into `TaskTable` instances via a `customColumns` array, as opposed to setting properties.  

In order to tackle this more granularly, we have elected to do this one column at a time. This is the first PR in this series of work and represents the injection of the docket number column.