Only update VACOLS location for ColocatedTasks when location is CASEFLOW (#10853)
There are 49 dispatched legacy appeals with active `ColocatedTask`s in our system (see #8810). When those `ColocatedTask`s are completed they will change the VACOLS location code to the location of the creator of the `ColocatedTask`. This PR changes that behaviour so Caseflow only changes the location of a case in VACOLS if the current location of that case is "CASEFLOW".