Automatic case distribution for Hearing Docket (#10704)
Resolves #9999

In order to support distribution of appeals in the Hearing docket, we
needed to define the `age_of_n_oldest_priority_appeals` and
`distribute_appeals` methods to adhere to the conditions that are
specific to Hearing appeals.

In `age_of_n_oldest_priority_appeals`, the first set of base conditions
that get applied are: priority, ready for distribution, and in the
hearing docket. These scopes (defined in the `Appeal` model) get
applied via the `Docket` class. An appeal is considered prioritized if
it is Advanced on the Docket. More details about priority status here:
https://github.com/department-of-veterans-affairs/caseflow/wiki/Automatic-Case-Distribution#priority

For appeals in the Hearing docket, we only want `genpop` appeals, which
are those that adhere to any of these conditions:
- with no hearings
- with hearings, but none that are held
- where the most recent held hearing was held by an inactive judge

This means that if an appeal has multiple hearings, one or more of
which have the `disposition` attribute set to `held`, we only want to
look at the most recent hearing. In order to determine the date the
hearing was held, we need to join on the `HearingDay` table and look
at the `scheduled_for` field.

More background info on `genpop` here:
https://github.com/department-of-veterans-affairs/caseflow/wiki/Automatic-Case-Distribution#genpop

For the `distribute_appeals` method, the base conditions are: ready for
distribution and in the hearing docket. Priority and genpop are passed
in via `AmaCaseDistribution#ama_distribution` because both priority and
nonpriority cases can be distributed, as well as "only genpop",
"not genpop", or "any" (which is a combination of the previous two).