Move LOB redirect from tasks.json endpoint to main .html endpoint (#10879)
This reverts 6b3a1e35625966efde4b203bf6e1d19e61d5f0da which put a redirect on the Organization tasks endpoint. That endpoint was only accessed via XHR for JSON. The redirect properly belongs on the /organizations/:name endpoint instead, which is the main .html response to the browser that triggers the XHR JSON request.