Show withdrawn status for request issues
If a request issue has been withdrawn, we want that to be apparent in
the case details view. We also want to be able to create decision
issues for withdrawn request issues. For a better UX, we automatically
set the decision issue disposition to "Withdrawn" in the dropdown if
the request issue is withdrawn, but we allow the user to change it if
necessary.