Refactor BvaDispatchTask#outcode
**Why**: Issue #10480 requires us to implement a new feature that only
applies to AMA appeals. The current code is not structured in such a
way that makes adding the feature easy. Keeping the existing code
would increase the complexity of the `outcode` method and reduce
readability.

In order to make adding the new feature easy without adding complexity,
we must first perform this preparatory refactoring.

The `outcode` method contains the "Switch Statements" code
smell which performs different actions based on the appeal's class.
The refactoring recipe we will use to remove this code smell is
"Replace conditional with polymorphism". We will adhere to the
"Tell, Don't Ask principle" and create separate classes for dispatching
AMA appeals and Legacy Appeals.

By extracting separate classes, we can also move the temporary `task`
and `tasks` variables into reusable methods. This recipe is known as
"Replace Temp with Query".

I also replaced actions such as `task.update!(status: "completed")`
with more descriptive method names like `complete_dispatch_task!`.

Note that we still have a conditional based on class in `outcode`. The
way to eliminate that conditional would be to have an `outcode` or
`bva_dispatch` method defined in the `Appeal` and `LegacyAppeal`
models that then call `AmaAppealDispatch` and `LegacyAppealDispatch`
respectively. Because those models are already too big, adding more
code would result in a Code Climate error. Since we don't anticipate
having more appeal types in the near future, I think this conditional
is fine for now.

References:
https://martinfowler.com/articles/preparatory-refactoring-example.html
https://refactoring.guru/smells/switch-statements
https://refactoring.guru/replace-conditional-with-polymorphism
https://refactoring.guru/replace-temp-with-query