Move ACD scopes from Appeal to Docket
**Why**: These scopes only apply to automatic case distribution, and
some aren't used alone. For example, we would never call
`Appeal.ready_for_distribution` in the app because we check appeals
for each docket separately. `ready_for_distribution` is an additonal
scope applied to a base query that checks for active appeals with
a particular docket type. Therefore, it doesn't make sense to place
these scopes in the Appeal class and test them individually since
they aren't used that way.

For now, the logical place to put these scopes is in the Docket class.
Once we continue refactoring the ACD code, we might find a better
place.