Refactor nonpriority_receipts_per_year
**Why**: To make it easier to understand.

**How**: Use the Replace Temp with Query refactoring recipe to move
all the variables and logic into private methods so that the main
method is understandable at a glance. The well-named methods explain
what is returned, and if one needs to dig deeper, they can look at the
private methods, but it's no longer necessary to read all of that code
in order to understand what the method is doing.

References:
https://refactoring.guru/replace-temp-with-query