Add efolder_documents_fetched_at to users table (#10979)
### Description
Reader job was recently changed to run for users who are active within a week that means we sould have to delete and re-create records in ReaderUser table. ReaderUser model appears to be over-loaded with unnecessary logic. Move efolder_documents_fetched_at to the users table to simplify the logic and code around it. 

### To do:
Delete reader_users table and all the code around it after this code is deployed

*Schema Changes Only*

* [x] Column comments updated
* [x] Query profiling performed (eyeball Rails log, check bullet and fasterer output)
* [x] Appropriate indexes added (especially for foreign keys, polymorphic columns, and unique constraints)
* [x] #appeals-schema notified with summary and link to this PR