Organization clean_url means voc_rehab -> voc-rehab (#10988)
**Why**

E.g. previously we did a `find` using the `url` "voc_rehab". Finding zero results, we tried to create it using "voc-rehab" because the `clean_url` before_save hook in Organization model altered the string. That means we would try to insert a duplicate organization.

This change assumes every BusinessLine type Organization has a unique name and that is sufficient for the initial `find`.