Allow withdrawn appeals to be distributed
**Why**: Judges need to provide a decision for withdrawn appeals.
The definition of a withdrawn appeal is one where all request issues
have been withdrawn, even if they are ineligible or closed for other
reasons.

Previously, withdrawing all request issues would cancel all of the
appeal's active tasks. However, in order to allow the appeal to be
distributed, we cannot cancel its `RootTask` or `TrackVeteranTask`,
and we need to set its `DistributionTask`'s status to `assigned`.

If an appeal has a mixture of withdrawn and ineligible request issues,
or a mixture of withdrawn and closed request issues, then we should
cancel all active tasks

Resolves #10973