Re-enable BGS lookup for POA (#11143)
This reverts commit ac9bf0016e24e9a9b841e85de7d24889b1e1df90.

see #11126 

When initially deployed, the BGS lookup was (on aggregate) too slow for production use. This re-enables the BGS lookup with 24 hours of Rails caching per-participant. There is also a job (which will eventually be run nightly) that can pre-warm all the caches so that users do not suffer the caching expense.