Update Appeal#withdrawn? to match latest requirements
While working on #11138, we understood the correct definition of a
withdrawn appeal, which is that it has at least one withdrawn request
issue, and no other active request issues. In most cases, when all of
an appeal's request issues are closed, its RootTask gets cancelled,
which is why it made sense for `withdrawn?` to check the RootTask
status. However, in the case of a withdrawal, the RootTask no longer
gets cancelled after implementing #11138.