Close EP400 request issues (#11215)
We have confirmation from AMO that if an end product gets its claim type changed to an EP400, that we can close the request issues with no_decision.  Currently we are doing this manually case-by-case, so that the request issues can be intaken again on a new decision review.
https://dsva.slack.com/archives/CFBLH2LR4/p1561396691147800

I have started looking into doing this for EPs that are already cleared. Since we don't store the claim_type_code if it gets changed, it will be good to prevent new incoming ones so that we don't have to search for new ones, which requires fetching the EPE result from BGS and takes some time.

Note, we do establish sometimes establish 400 claims from Caseflow as well, but those do not have any request issues.