Create report for catching untracked LegacyAppeals (#11258)
Connects #11183. This PR adds a job that sends a report to Slack when it runs if there are `LegacyAppeal`s without active Caseflow tasks charged to the "CASEFLOW" location in VACOLS. A follow-on PR will be created in the appeals-deployment repo after this is merged to run this report daily.