Allow enabling all features currently in use (#11274)
**Why**: Previously, the only way to ensure that one could easily
enable all feature toggles that were currently in use by the app was to
keep the config in `enable_features_dev.rb` up to date, which was rarely
the case.

Instead of relying on humans to keep a file up to date, I thought I'd
write a script that would grep the app for feature toggles, then
enable them all. This ensures that we are always using the features
used by the app, no more, no less.