Simplify and unify error handling
**Why**: In order to make it easier to handle errors from VBMS and BGS,
we moved the logic to those gems which now provide a handy `transient?`
method we can use to determine whether or not to send the error to
Sentry.

By adding `rescue_from` clauses in `ApplicationJob` and the parent
controllers, we ensure all transient errors can be caught.

There are likely still cases where we rescue from VBMS and BGS errors
directly in other files, but to keep this PR as small as possible, those
can be addressed later if we find we are still seeing transient errors
in Sentry.