Speed up test suite
**How**: Our biggest bottleneck in specs is the VACOLS-related DB
cleaning in `database_cleaner.rb`. Since only specific tests need to
interact with VACOLS, it makes sense to only perform the DB cleaning
for those tests.

- Create separate `database_cleaner` and `vacols_database_cleaner`
files that only clean the Postgres and Oracle DB, respectively.
- Don't automatically require these 2 files in `rails_helper`. Instead,
require one or the other as needed in each spec.
- When the entire suite is run, once it encounters a `require` statement
to load `database_cleaner` or `vacols_database_cleaner`, it will run the
DB cleaning in both files for all subsequent tests. To avoid that, we
additionally specify that the Postgres cleaning should only happen in
specs that are tagged with `:postgres`, and the Oracle cleaning in tests
tagged with `:vacols`.
- Move the loading of the `Vftypes`, `Issref`, and `Actcode` VACOLS
tables out of a `before(:all)` block and into a separate rake task
since this only needs to run once for the lifetime of the test DB. We
don't want to put it in a `before(:suite)` block because that will slow
down running individual tests locally.
- Move other helper methods out of `rails_helper` into separate files
for better organization