Migrate created_by and updated_by values on HearingDays to associations (#11456)
Connects #11260. This is the second of several PRs that'll resolve that issue; the first was #11435.

### Description
Adds a rake task with tests that examines each `HearingDay` object and transforms the css_ids it finds in the `created_by` and `updated_by` fields into IDs for the correct user in the `created_by_id` and `updated_by_id` fields.

After this PR is merged and on production, I'll follow this script:

1. Dry run the task. This'll surface any irregularities (`css_id`s that don't correspond to users in the database, etc.) that can be addressed before doing the actual migration:
  ```bundle exec rake data_migrations:migrate_hearing_day_created_and_updated_by```

1. Address the irregularities discovered in the dry run.

1. Run the task to perform the migration. The task's output will be preserved in the application logs and in stdout.
  ```bundle exec rake data_migrations:migrate_hearing_day_created_and_updated_by[false]```

The task and its tests will be removed in the final PR connected to this issue.