Do not check for duplicate issue by rating_issue_reference_id if nil (#11528)
connects #11083 

The eligibility check was marking issues as duplicate by rating_issue_reference_id if it was from a decision issue, but there was no duplicate decision issue and the rating issue reference id on both issues was nil.

What do we do for clean up on this?  We can't make the issues eligible again.  I guess to start it would be good to check how many issues are impacted.