Defensively cancel TaskTimers with closed parents during job runs (#11591)
Bumps #11442 

### Description
We now cancel TaskTimers on Tasks that are cancelled through a callback (#11581).  

This additionally adds a defensive cancel if we find TaskTimers with closed parents when the job runs.

This should also close any outstanding TaskTimers that were orphaned before #11581 was released on it's first run post release. Confirming that happens properly is the last step in #11442 before it can be marked completed.