Refactor upload_document_to_vbms to service class
**Why**: To separate the concerns and have each related VBMS call
defined in its own class. This follows the pattern we started by moving
document uploads to the `ExternalApi::VbmsDocumentsForAppeal` class.

The main reason for this refactor is to be able to try a different
file number (using the BGS claim number) if the veteran's file number
does not exist in VBMS. We were not being consistent with trying this
alternative number.

Ulimately, we want to find a better long-term solution for trying
alternative numbers for any VBMS call where the file number is required.