Add user-friendly validation to `outcode` endpoint
**Why**: Previously, we were handling some validation errors by raising
exceptions and then rescuing them in various controllers in order to
respond with the appropriate JSON. This is not optimal for various
reasons:
- Raising errors is slow
- Exceptions should not be raised for flow of control
  (https://rubystyle.guide/#no-exceptional-flows)
- Exceptions should not be raised for expected scenarios, such as
invalid data from the user
- Leaving it up to Rails to raise exceptions makes it hard to see what
exactly is being validated, and results in inconsistent behavior and
bugs. For example, in some cases, we were returning JSON such as
`{ message: "error message" }`. In other cases, we were returning a
`Caseflow::Error`, which uses a different hash.

One common Sentry error we have been getting is
`ActiveRecord::RecordNotUnique` due to IDT trying to outcode an appeal
with an existing citation number (tied to a different appeal). This
error was not being rescued, and the controller was returning a 200
response nonetheless, which is misleading.

**How**:
- Use our pattern where the controller calls a class that performs
ActiveModel validations, and returns a `FormResponse` object that
responds to `success?` and `errors` so the controller can decide how
to respond.
- All the validations are now in one place and easily understandable
using the Rails validation conventions.
- No more raising exceptions for validation errors, making the
controller faster

Resolves #11354