pass ids to fixes:activate_stalled_hearing_tasks (#11662)
Connects #11070 

### Description
Enables passing of ids to the `fixes:activate_stalled_hearing_tasks` rake task, so that we can fix just a few tasks before we run the script to fix every task. The `limit` parameter does something similar, but still examines every `HearingTask` in the database to determine whether it's stalled or not.