Sort by veteran name (#11676)
* Add veteran_name to the cached attributes job

* Sort by veteran name on the back end

* Use quicker access to veteran names

* Fix n+1 query for Appeals

* Default sort order is by created_at

CASE_DETAILS_LINK_COLUMN has its own sorting case. For the default
sorting to drop through to the default `created_at`, the default sort_by
should be nil

* Fix RegionalOffice::ROS.shuffle to be truely uniq

There are multiple regional offices with a city of "Washington". This
causes a flakey test when multiple tasks receive the same city, which
results in an expected sort order mismatch

* Ignore false positive code climate

[ci skip]

* Fix n+1 queries for LegacyAppeals

* Add sneaky column comment

* PR review comments